ステージデータ取得用インターフェース実装
