--- Assets/Script/GameScene/Interfaces/IJsonFileManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bf5bc50f6117dc44f89f9aed7426fd05
