--- Assets/Script/GameScene/JsonFileManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 679dd047149cc0641a5bf5f1f1f7bedc
