+++ Assets/Script/General/Interfaces/IJsonFileManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 569132cbb1efb2c4b80c83ed82ceb4cc
