+++ Assets/Script/General/JsonFileManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6e89d4aea01b9a84494c436df255b2fa
